Changed deck0.tex so that it uses definitions_default.tex to define some commands.  See a previous commit.
I also added notes and an option for compiling the tex with just the
notes.

In a few places I inserted alternate text.  The locations of the
alternate text are makred by
% ALT ALT
so that one can easilly search for those changes.

The alternative text has been commented out so that Mine's original work
is unchanged.